+++ mayan/apps/quotas/forms.py
@@ -33,11 +33,7 @@ class QuotaDynamicForm(forms.DynamicModelForm):
         )
         result = super().__init__(*args, **kwargs)
 
-        # Handle filtered queryset fields.
-        for field in self.fields:
-            if hasattr(self.fields[field], 'reload'):
-                self.fields[field].user = self.user
-                self.fields[field].reload()
+        self.do_fields_reload()
 
         if self.instance.backend_data:
             for key, value in json.loads(s=self.instance.backend_data).items():
@@ -70,3 +66,6 @@ class QuotaDynamicForm(forms.DynamicModelForm):
 
         data['backend_data'] = json.dumps(obj=backend_data)
         return data
+
+    def get_field_reload_attributes(self):
+        return {'user': self.user}