+++ HISTORY.rst
@@ -1,6 +1,8 @@
 4.9.6 (XXXX-XX-XX)
 ==================
 - Rename file metadata model mixin class for clarity.
+- Consolidate dynamic filtering field reload code. Add the new
+  `FormMixinFilteredFieldsReload` mixin.
 
 4.9.5 (2025-09-28)
 ==================